docs(demo): minor demo page cleanup and styling tweaks
